requirements.txt
# Requirements for running notebooks
-e .

# For running tests
nose

# For neural network examples
keras
tensorflow
torch
torchvision

# For decision tree examples
xgboost
catboost
lightgbm

# Server
Flask
google-cloud-storage
google-api-core
google-auth
google-cloud-core
